+++ src/client/java/no2/bugfixerupper/mixin/client/misc/leashed_player_vehicle/EntityMixin.java
@@ -1,8 +1,9 @@
 package no2.bugfixerupper.mixin.client.misc.leashed_player_vehicle;
 
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.Leashable;
-import net.minecraft.world.entity.vehicle.AbstractBoat;
 import net.minecraft.world.level.Level;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -31,25 +32,29 @@ public abstract class EntityMixin {
         }
     }
 
+    @SuppressWarnings("StatementWithEmptyBody")
     @Unique
-    private void clientLeashTick() {
-        Leashable entity = (Leashable) this;
-        Entity holder = entity.getLeashHolder();
-        if (holder != null) {
-            float distance = this.distanceTo(holder);
-
-            if (distance > 6f) {
-                //noinspection ConstantValue
-                if (this.onGround() && !(entity instanceof AbstractBoat)) { //AbstractBoat has its own smooth implementation of elasticRangeLeashBehavior
-                    if (distance < 6.25f) {
-                        //Smooth out the cutoff when on the ground to avoid glitchy / bouncy leads, e.g. when riding a leashed horse. When not on the ground, leads are supposed to be bouncy.
-                        double lerp = ((distance - 6.0f) * 4); //0 at 6.0, 1 at 6.25, i.e. 4 pixels in
-                        //Pull strength is divided by distance squared, by multiplying 1/sqrt(lerp), the result is multiplied by lerp
-                        //Using a linear lerp because this is minecart, which lerps so many things.
-                        distance *= (float) (1.0 / (Math.sqrt(lerp)));
-                    }
+    private <E extends Entity & Leashable> void clientLeashTick() {
+        //Vanilla code adapted for client side
+        //noinspection unchecked,DataFlowIssue
+        E entity = (E) (Object) this;
+        Leashable.LeashData leashData = entity.getLeashData();
+        if (leashData != null && leashData.leashHolder != null) {
+            Entity holder = entity.getLeashHolder();
+            if (holder != null && holder.level() == entity.level()) {
+                double d = entity.leashDistanceTo(holder);
+                entity.whenLeashedTo(holder);
+                if (d > entity.leashSnapDistance()) {
+                    //breaking leash, handled server side
+                } else if (d > entity.leashElasticDistance() - (double) holder.getBbWidth() - (double) entity.getBbWidth()
+                        && /*CLIENT SIDE RELEVANT PUSHING:*/ entity.checkElasticInteractions(holder, leashData)) {
+                    //fall damage + ai stuff, handled server side
+                } else {
+                    //pathfinding leash, handled server side
                 }
-                entity.elasticRangeLeashBehaviour(holder, distance);
+
+                entity.setYRot((float) ((double) entity.getYRot() - leashData.angularMomentum));
+                leashData.angularMomentum = leashData.angularMomentum * (double) Leashable.angularFriction(entity);
             }
         }
     }