+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 BugFixerUpper _ModVersion_ for Minecraft _MCVersion_ adds a new feature:
 
-## Additions:
-### Misc:
-- Ghast balls, wind charges and shulker bullets do not have an extended click box for creative mode players (Install on client!)
+## Changes
+- Adapt client side leash fix for new leash system (quad leashes, less bouncy leashes)